# 8274308: Improve efficiency for HandshakeContext initialization.