@djelinski 2021-10-03
I quickly hacked together a SSLEngine benchmark for server-side handshakes [here](https://github.com/djelinski/jdk/blob/2bee5c471b41d60d4248d5d16e868060f0efda19/test/micro/org/openjdk/bench/java/security/SSLEngineStartHandshake.java); note that on the server side I run the following code on every connection;
```
            SSLEngine serverEngine = context.createSSLEngine();
            serverEngine.setUseClientMode(false);
            serverEngine.setEnabledProtocols(new String[]{"TLSv1.3"});
            serverEngine.setEnabledCipherSuites(new String[]{"TLS_AES_256_GCM_SHA384"});
            serverEngine.beginHandshake();
```
not sure if that's the best way to initialize SSLEngine. Anyway, in this setup I observed a 16% performance degradation; results without your patch:
```
Benchmark                                         Mode  Cnt  Score   Error   Units
SSLEngineStartHandshake.handshakeBenchmark       thrpt   25  0.837 ± 0.006  ops/ms
SSLEngineStartHandshake.handshakeBenchmark        avgt   25  1.187 ± 0.006   ms/op
```
with your patch:
```
Benchmark                                         Mode  Cnt  Score   Error   Units
SSLEngineStartHandshake.handshakeBenchmark       thrpt   25  0.702 ± 0.009  ops/ms
SSLEngineStartHandshake.handshakeBenchmark        avgt   25  1.416 ± 0.013   ms/op
```