@cliveverghese 2021-10-03
Hi @djelinski 

Thank you for putting this together. I will investigate the results of your benchmarks.