@cliveverghese 2021-10-20
Thank you for the feedback. 

Based on the feedback provided here, I am working on moving this to a cache within HandshakeContext. I will be updating the PR with that implementation soon for your feedback.