@cliveverghese 2021-10-29
Hi, I have update the PR with an alternate implementation. 

The updated benchmark results are 

New Benchmarks 
```
Benchmark                                         Mode  Cnt  Score   Error   Units
SSLStartHandshake.handshakeBenchmark             thrpt   25  0.241 ± 0.016  ops/ms
SSLStartHandshake.handshakeBenchmark              avgt   25  4.010 ± 0.224   ms/op
SSLEngineStartHandshake.handshakeBenchmark       thrpt   25  0.642 ± 0.021  ops/ms
SSLEngineStartHandshake.handshakeBenchmark        avgt   25  1.552 ± 0.058   ms/op
```

Old Benchmarks
```
Benchmark                                         Mode  Cnt  Score   Error   Units
SSLStartHandshake.handshakeBenchmark             thrpt   25  0.229 ± 0.010  ops/ms
SSLStartHandshake.handshakeBenchmark              avgt   25  4.375 ± 0.283   ms/op
SSLEngineStartHandshake.handshakeBenchmark       thrpt   25  0.621 ± 0.035  ops/ms
SSLEngineStartHandshake.handshakeBenchmark        avgt   25  1.830 ± 0.406   ms/op
```

This new implementation uses a cache in the HandshakeContext class. 

Looking forward for you feedback.